Adjust for usage with OpenLDAP
We're changing the hard coded AD-specific stuff and changing it out for
usage with OpenLDAP by way of `inetOrgPerson` + `posixAccount` user
accounts, thus facilitating the need of using uid.

Additionally, the cache handler was switched to just using the user name,
which is a working solution for us.

Ideally, this should all be made configurable, but this is a "works good
enough" approach.